+++ jobbergate-cli/CHANGELOG.rst
@@ -9,6 +9,7 @@ Unreleased
 
 4.0.0a7 -- 2023-09-05
 ---------------------
+- Fixed supporting files output names
 
 4.0.0a6 -- 2023-08-31
 ---------------------